Lifecycle moved to Cluster class + optional Slack icon url
