+++ src/main/resources/component-web.conf
@@ -278,6 +278,9 @@ health {
         # Select the sender name here - leave empty to use the product name
         sender = ""
 
+        # Select the sender icon here - leave empty to use Slack default icon
+        icon_url = ""
+
         # Determines which message type to send to hip chat. Types are:
         #    start    - Start of the node
         #    stop     - Stop of a node
